# internal/controller/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Distributed Encoder API",
    "version": "1.0.0",
    "description": "REST API for the distributed video encoding controller."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/webhooks": {
      "get": {
        "summary": "List webhooks",
        "operationId": "listWebhooks",
        "tags": ["webhooks"],
        "responses": {
          "200": {
            "description": "List of configured webhooks"
          }
        }
      },
      "post": {
        "summary": "Create a webhook",
        "operationId": "createWebhook",
        "tags": ["webhooks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "provider", "url", "events"],
                "properties": {
                  "name": { "type": "string" },
                  "provider": { "type": "string", "enum": ["discord", "teams", "slack"] },
                  "url": { "type": "string", "format": "uri" },
                  "events": { "type": "array", "items": { "type": "string" } },
                  "secret": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook created"
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457 Problem Details",
        "required": ["type", "title", "status"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "URI reference identifying the problem type"
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference",
            "description": "URI reference identifying the specific occurrence"
          },
          "request_id": {
            "type": "string",
            "description": "Correlation identifier for the request"
          }
        }
      }
    }
  }
}